Add files via upload
In this file ("list.json") are saved basic server information. Specifically, its IPv4 address, MAC address (needed to construct the Wake on LAN packet), WOL port and minecraft server ports. The "list.json" file will be created by the program itself if it is not found in the same folder where the program itself is located. The file can be edited manually, or via the program, where there is an "editor" function that reads the information and converts it into the appropriate format.